busy and booked days apis, change create album

+++ webapi/src/main/java/fpt/university/pbswebapi/repository/AlbumRepository.java
@@ -15,5 +15,5 @@ public interface AlbumRepository extends JpaRepository<Album, Long> {
     Page<Album> findAllSortByLike(Pageable paging);
 
     @Query("from Album album where album.category.id =:categoryId order by album.likes desc ")
-    Page<Album> findByCategoryIdSortByLike(Pageable paging, int categoryId);
+    Page<Album>     findByCategoryIdSortByLike(Pageable paging, long categoryId);
 }
